utils/requirements.txt
watchdog==3.0.0
Pillow==10.1.0
numpy==1.24.3 